Restrict manipulating not own sections/threads/replies

+++ src/backend/LDW.Application/Features/ForumFeatures/Commands/DeleteForumSectionCommand.cs
@@ -1,7 +1,9 @@
 ﻿using LDW.Application.Interfaces;
 using LDW.Domain.Common.Exceptions;
 using MediatR;
+using Microsoft.AspNetCore.Http;
 using System;
+using System.Security.Claims;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -14,18 +16,27 @@ namespace LDW.Application.Features.ForumFeatures.Commands
 		public class DeleteForumSectionCommandHandler : IRequestHandler<DeleteForumSectionCommand, Guid>
 		{
 			private readonly IApplicationDbContext _context;
-			public DeleteForumSectionCommandHandler(IApplicationDbContext context)
+			private readonly IHttpContextAccessor _httpContextAccessor;
+
+			public DeleteForumSectionCommandHandler(IApplicationDbContext context, IHttpContextAccessor httpContextAccessor)
 			{
 				_context = context;
+				_httpContextAccessor = httpContextAccessor;
 			}
 
 			public async Task<Guid> Handle(DeleteForumSectionCommand request, CancellationToken cancellationToken)
 			{
 				var forumSectionToDelete = await _context.ForumSections.FindAsync(request.Id);
+				var currentLoggedInUserId = _httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier);
 
 				if (forumSectionToDelete == null)
 				{
 					throw new NotFoundException("ForumSection", request.Id);
+				}
+
+				if (forumSectionToDelete.AuthorId != currentLoggedInUserId)
+				{
+					throw new AccessForbiddenException("ForumSection", request.Id);
 				}
 
 				_context.ForumSections.Remove(forumSectionToDelete);
